+++ PowerPointConverter/Model/HtmlConvertInfo.cs
@@ -4,8 +4,8 @@
     {
         public int Index { get; set; }
         public int Number { get; set; }
-        public decimal Width { get; set; }
-        public decimal Height { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public string Html { get; set; }
         public bool IsOK { get; set; }
         public string Message { get; set; }
